Changed requested things, moved util method, etc
